replace old version of function module
