+++ dev/v1.ex
@@ -1,7 +1,7 @@
 defmodule Inngest.Dev.EventFn2 do
   @moduledoc false
 
-  use Inngest.V1.Function
+  use Inngest.Function
   alias Inngest.Function.{Opts, Trigger}
 
   @func %Opts{id: "test-func-v2", name: "test func v2"}
