+++ lib/inngest/function/opts.ex
@@ -1,0 +1,13 @@
+defmodule Inngest.Function.Opts do
+  defstruct [
+    :id,
+    :name,
+    :retries
+  ]
+
+  @type t() :: %__MODULE__{
+          id: binary(),
+          name: binary(),
+          retries: number()
+        }
+end
